Squashing alembic revisions
